+++ changelog.md
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+0.11.0
+------
+
+- #52 cleanup model registry
+- #10 added speech to text using silero_vad and faster_whisper
+
 0.10.0
 ------